Update script.js
In this commit, I add some prompts for the user to input. To keep the code cleaner, I'm creating a separate function to hold my prompts.  These prompts are directly tied to the arrays I've established above. I'm only missing one prompt and that's the character length prompt. I'm still working on that prompt. When it is finished, I will put it in my ''prompts'' function.